Make `useCookie` hook universal
