+++ public/universal/hooks.ts
@@ -1,7 +1,11 @@
+import type { CookieSerializeOptions } from 'cookie';
+
 import { type RootTransitionOptions, usePageTransition } from '@/private/client/hooks';
+import { RootServerContext } from '@/private/server/context';
 import { usePrivateLocation, useUniversalContext } from '@/private/universal/hooks';
 import type { CustomNavigator } from '@/private/universal/types/util';
 import { populatePathSegments } from '@/private/universal/utils/paths';
+import { useContext } from 'react';
 
 const useParams = <
   TParams extends Record<string, unknown> | Array<string> = Record<
@@ -60,11 +64,11 @@ interface RedirectOptions
 }
 
 const useRedirect = () => {
+  const populatePathname = usePopulatePathname();
+
   // @ts-expect-error The `Netlify` global only exists in the Netlify environment.
   const isNetlify = typeof Netlify !== 'undefined';
   if (typeof window === 'undefined' || isNetlify) {
-    const populatePathname = usePopulatePathname();
-
     return (pathname: string, options?: Pick<RedirectOptions, 'extraParams'>): never => {
       const populatedPathname = populatePathname(pathname, options?.extraParams);
 
@@ -73,7 +77,6 @@ const useRedirect = () => {
   }
 
   const transitionPage = usePageTransition();
-  const populatePathname = usePopulatePathname();
 
   return (pathname: string, options?: RedirectOptions) => {
     const populatedPathname = populatePathname(pathname, options?.extraParams);
@@ -84,4 +87,68 @@ const useRedirect = () => {
   };
 };
 
-export { useLocation, useNavigator, useParams, usePopulatePathname, useRedirect };
+export type CookieHookOptions = {
+  /**
+   * Allows for making cookies accessible to the client, instead of allowing only the
+   * server to read and modify them.
+   */
+  client?: true;
+  /**
+   * Allows for restricting the cookie to a specific URL path of the app.
+   *
+   * @default '/'
+   */
+  path?: string;
+};
+
+const useCookie = <T extends string | null>(
+  name: string,
+): [T | null, (value: T, options?: CookieHookOptions) => void] => {
+  const serverContext = useContext(RootServerContext);
+  if (!serverContext) throw new Error('Missing server context in `useCookie`');
+
+  const { cookies, collected } = serverContext;
+  const value = cookies[name] as T | null;
+
+  const setValue = (value: T, options?: CookieHookOptions) => {
+    const cookieSettings: CookieSerializeOptions = {
+      // 365 days.
+      maxAge: 31536000,
+      httpOnly: !options?.client,
+      path: options?.path || '/',
+    };
+
+    // To delete cookies, we have to set their expiration time to the past.
+    if (value === null) {
+      cookieSettings.expires = new Date(Date.now() - 1000000);
+      delete cookieSettings.maxAge;
+    }
+    // As per the types defined for the surrounding function, this condition would never
+    // be met. But we'd like to keep it regardless, to catch cases where the types
+    // provided to the developer aren't smart enough to avoid `undefined` or similar
+    // getting passed.
+    else if (typeof value !== 'string') {
+      let message = 'Cookies can only be set to a string value, or `null` ';
+      message += `for deleting them. Please adjust "${name}".`;
+      throw new Error(message);
+    }
+
+    if (!collected.cookies) collected.cookies = {};
+
+    collected.cookies[name] = {
+      value,
+      ...cookieSettings,
+    };
+  };
+
+  return [value, setValue];
+};
+
+export {
+  useLocation,
+  useNavigator,
+  useParams,
+  usePopulatePathname,
+  useRedirect,
+  useCookie,
+};
